major update. Webpack optimization and support of multiple webworkers
