+++ package.json
@@ -6,10 +6,10 @@
     "DEX"
   ],
   "scripts": {
-    "build-deploy": "webpack --config build/deploy/webpack-deploy.config.js",
-    "build-wallet": "webpack --config build/webpack-wallet.config.js",
-    "build-wallet-analyzer": "webpack --config build/webpack-wallet.config.js --mode development --analyzer",
-    "build-pandorapay-webworker": "webpack --config build/webpack-pandorapay-websocket.config.js",
+    "build-wallet": "webpack --config build/webpack-wallet.config.js --mode production --production",
+    "build-wallet-analyzer": "webpack --config build/webpack-wallet.config.js --mode production --production --analyzer",
+    "build-pandorapay-webworker": "webpack --config build/webpack-pandorapay-websocket.config.js --mode production --production",
+    "build-pandorapay-helper-webworker": "webpack --config build/webpack-pandorapay-helper-websocket.config.js --mode production --production",
     "build": "npm run build-wallet",
     "test": "npm run build-wallet",
     "dev": "webpack-dev-server --config=./build/webpack-hot-loader.config.js --content-base dist/dev --hot --inline --host 0.0.0.0"
