+++ src/intro-app/pandorapay-webworker-integration.js
@@ -1,11 +1,82 @@
 import consts from "consts/consts";
-import Helper from "src/webworkers/pandorapay-webworker/helper"
+import Helper from "src/webworkers/helpers/helper"
 
 export default class PandorapayWebworkerIntegration{
 
-    constructor() {
+    constructor(name, wasmFileName, workerFileName, initializeStatusEvent, initializedEvent) {
+        if (!name) throw "name was not defined"
+        this.name = name
+        this.wasmFileName = wasmFileName
+        this.workerFileName = workerFileName
 
-        this.worker = new Worker(PandoraPayWalletOptions.resPrefix+"PandoraPay-webworker.js");
+        this.initializeStatusEvent = initializeStatusEvent || function () {}
+        this.initializedEvent = initializedEvent || function (){}
+    }
+
+    async downloadWasm(progressStatusCallback){
+
+        const response = await fetch(PandoraPayWalletOptions.resPrefix+this.wasmFileName, {
+            headers: {
+                'Content-Encoding': 'gzip',
+            }
+        })
+
+        if (!response.ok) throw response.status+' '+response.statusText
+        if (!response.body) throw 'ReadableStream not yet supported in this browser.'
+
+        // to access headers, server must send CORS header "Access-Control-Expose-Headers: content-encoding, content-length x-file-size"
+        // server must send custom x-file-size header if gzip or other content-encoding is used
+        const contentEncoding = response.headers.get('content-encoding');
+        let contentLength = response.headers.get( contentEncoding ? 'x-file-size' : 'content-length');
+        if (contentLength === null && contentEncoding) {
+            console.error('Response size header unavailable for encoded');
+            contentLength = response.headers.get('content-length');
+        }
+
+        if (contentLength === null) throw "Response size header unavailable"
+
+        const total = parseInt(contentLength, 10);
+        let loaded = 0;
+
+        let r = await ( new Response(
+                new ReadableStream({
+                    start(controller) {
+                        const reader = response.body.getReader();
+
+                        let lastTransferred = 0
+                        function read() {
+                            reader.read().then(({done, value}) => {
+                                if (done) {
+                                    controller.close();
+                                    return;
+                                }
+                                loaded += value.byteLength;
+
+                                if (loaded - lastTransferred > 20240) {
+                                    lastTransferred = loaded
+                                    progressStatusCallback(loaded, total)
+                                }
+
+                                controller.enqueue(value);
+                                read();
+
+                            }).catch(error => {
+                                controller.error(error)
+                            })
+                        }
+
+                        read();
+
+                    }
+                })
+        ) );
+
+        return r
+    }
+
+    createWorker(){
+
+        this.worker = new Worker(PandoraPayWalletOptions.resPrefix+this.workerFileName);
 
         this.worker.onmessage = async (event) => {
 
@@ -15,20 +86,15 @@ export default class PandorapayWebworkerIntegration{
                 //console.log(data)
 
                 if (data.type === "initialize-answer") {
-                    if (this.initializeStatusEvent) this.initializeStatusEvent(data.status)
+                    this.initializeStatusEvent(data.status)
                 }
                 else if (data.type === "initialize-done") {
 
                     const {PandoraPayClone} = data
 
-                    this.fixPandoraPay(PandoraPayClone)
+                    this.fixInstanceObject(PandoraPayClone)
 
-                    await PandoraPay.helpers.helloPandora()
-
-                    if (this.initializeStatusEvent) this.initializeStatusEvent("PandoraPay WASM is working!")
-
-                    PandoraPayWallet.loadWallet()
-
+                    this.initializedEvent()
                 }
 
                 return Helper.OnMessage(this.worker, data)
@@ -46,9 +112,9 @@ export default class PandorapayWebworkerIntegration{
         });
     }
 
-    fixPandoraPay(src){
+    fixInstanceObject(src){
         const dst = Helper.ProcessObject(src, [] )
-        global.PandoraPay = Helper.FixObject(this.worker, dst)
+        global[this.name] = Helper.FixObject(this.worker, dst)
     }
 
 }
