+++ src/webworkers/helpers/storage/browser-storage.js
