+++ src/webworkers/pandorapay-helper-webworker/pandorapay-helper-webworker.js
@@ -1,0 +1,2 @@
+const GoWebworker = require('../helpers/go-webworker')
+
